+++ javascript/modules/client/src/client-addressList.ts
@@ -31,6 +31,7 @@ import {
   VoteProposalStepValue,
 } from "./internal/interfaces/plugins";
 import {
+  delay,
   getDummyAddressListProposal,
   getDummyAddressListProposalListItem,
 } from "./internal/temp-mock";
@@ -228,12 +229,13 @@ export class ClientAddressList extends ClientCore
     }
 
     // TODO: Implement
-
+    await delay(1000);
     yield {
       key: ProposalCreationSteps.CREATING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
+    await delay(3000);
     yield {
       key: ProposalCreationSteps.DONE,
       proposalId:
@@ -252,12 +254,13 @@ export class ClientAddressList extends ClientCore
     }
 
     // TODO: Implement
-
+    await delay(1000);
     yield {
       key: VoteProposalStep.VOTING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
+    await delay(3000);
     yield {
       key: VoteProposalStep.DONE,
       voteId:
@@ -276,12 +279,13 @@ export class ClientAddressList extends ClientCore
     }
 
     // TODO: Implement
-
+    await delay(1000);
     yield {
       key: ExecuteProposalStep.EXECUTING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
+    await delay(3000);
     yield {
       key: ExecuteProposalStep.DONE,
     };