+++ javascript/modules/client/src/client-erc20.ts
@@ -35,6 +35,7 @@ import {
 } from "./internal/encoding/plugins";
 import { bytesToHex, Random } from "@aragon/sdk-common";
 import {
+  delay,
   getDummyErc20Proposal,
   getDummyErc20ProposalListItem,
 } from "./internal/temp-mock";
@@ -239,13 +240,13 @@ export class ClientErc20 extends ClientCore implements IClientErc20 {
     }
 
     // TODO: Remove below as the new contracts are ready
-
+    await delay(1000);
     yield {
       key: ProposalCreationSteps.CREATING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
-
+    await delay(3000);
     yield {
       key: ProposalCreationSteps.DONE,
       proposalId:
@@ -290,12 +291,13 @@ export class ClientErc20 extends ClientCore implements IClientErc20 {
     }
 
     // TODO: Implement
-
+    await delay(1000);
     yield {
       key: VoteProposalStep.VOTING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
+    await delay(3000);
     yield {
       key: VoteProposalStep.DONE,
       voteId:
@@ -314,12 +316,13 @@ export class ClientErc20 extends ClientCore implements IClientErc20 {
     }
 
     // TODO: Implement
-
+    await delay(1000);
     yield {
       key: ExecuteProposalStep.EXECUTING,
       txHash:
         "0x0123456789012345678901234567890123456789012345678901234567890123",
     };
+    await delay(3000);
     yield {
       key: ExecuteProposalStep.DONE,
     };
@@ -331,7 +334,7 @@ export class ClientErc20 extends ClientCore implements IClientErc20 {
     params: IPluginSettings,
   ): DaoAction {
     if (!isAddress(pluginAddress)) {
-      throw new Error("Invalid plugin address")
+      throw new Error("Invalid plugin address");
     }
     // TODO: check if to and value are correct
     return {